refactor use Auth hook
